# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-circlecircumference",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute the circumference of a circle, `2 * pi * radius`.",
        "description": "This service owns the *control flow* but delegates every arithmetic step to\nits dependencies, exactly as specified:\n\n1. ask the constant service `srvcs-pi` (with an EMPTY body) for `p = pi`;\n2. ask `srvcs-floatmultiply` for `d = 2 * radius` (the diameter);\n3. ask `srvcs-floatmultiply` again for `result = p * d` — i.e.\n   `circumference = pi * (2 * radius)`.\n\nThis service never calls `srvcs-isnumber` directly: `srvcs-floatmultiply`\nvalidates its operands and any `422` it raises is forwarded verbatim. If a\ndependency is unreachable it reports itself degraded (`503`).",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CircumferenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "a dependency rejected the input (forwarded)"
          },
          "500": {
            "description": "a dependency returned a malformed result"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CircumferenceResponse": {
        "type": "object",
        "required": [
          "radius",
          "result"
        ],
        "properties": {
          "radius": {
            "type": "object"
          },
          "result": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "EvalRequest": {
        "type": "object",
        "required": [
          "radius"
        ],
        "properties": {
          "radius": {
            "type": "object",
            "description": "The radius of the circle."
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      }
    }
  }
}